Move interfaces into their own folders
