+++ src/EPR.SubsidiaryBulkUpload.Application.UnitTests/Services/CompaniesHouseDataProviderTests.cs
@@ -2,6 +2,7 @@
 using EPR.SubsidiaryBulkUpload.Application.Models;
 using EPR.SubsidiaryBulkUpload.Application.Options;
 using EPR.SubsidiaryBulkUpload.Application.Services;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 using Microsoft.Extensions.Options;
 
 namespace EPR.SubsidiaryBulkUpload.Application.UnitTests.Services;
