+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDataProvider.cs
@@ -1,5 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Models;
 using EPR.SubsidiaryBulkUpload.Application.Options;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 using Microsoft.Extensions.Options;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services;
