+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/CompaniesHouseWebCrawlerService.cs
@@ -1,5 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Exceptions;
 using EPR.SubsidiaryBulkUpload.Application.Extensions;
+using EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces;
 using Microsoft.Extensions.Logging;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload;
