+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/FileDownloadService.cs
@@ -1,4 +1,5 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Models;
+using EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces;
 using Microsoft.Extensions.Logging;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload;
