+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/HtmlWebProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+using EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces;
 using HtmlAgilityPack;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload;
