+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/Interfaces/ICompaniesHouseDownloadService.cs
@@ -1,4 +1,4 @@
-﻿namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload;
+﻿namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces;
 
 public interface ICompaniesHouseDownloadService
 {
