+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/Interfaces/ICompaniesHouseWebCrawlerService.cs
@@ -1,4 +1,4 @@
-﻿namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload
+﻿namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces
 {
     public interface ICompaniesHouseWebCrawlerService
     {
