+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/Interfaces/IDownloadStatusStorage.cs
@@ -1,6 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces;
 
 public interface IDownloadStatusStorage
 {
