+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseDownload/Interfaces/IFileDownloadService.cs
@@ -1,6 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload
+namespace EPR.SubsidiaryBulkUpload.Application.Services.CompaniesHouseDownload.Interfaces
 {
     public interface IFileDownloadService
     {
