+++ src/EPR.SubsidiaryBulkUpload.Application/Services/CompaniesHouseLookupDirectService.cs
@@ -3,6 +3,7 @@ using System.Text.Json;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Extensions;
 using EPR.SubsidiaryBulkUpload.Application.Models;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services;
