+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/IBulkSubsidiaryProcessor.cs
@@ -1,7 +1,7 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 public interface IBulkSubsidiaryProcessor
 {
