+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/ICsvProcessor.cs
@@ -1,6 +1,6 @@
 ﻿using CsvHelper.Configuration;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 public interface ICsvProcessor
 {
