+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/IParserClass.cs
@@ -1,7 +1,7 @@
 ﻿using CsvHelper.Configuration;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces
 {
     public interface IParserClass
     {
