+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/IRecordExtraction.cs
@@ -1,6 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.DTOs;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 public interface IRecordExtraction
 {
