+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/ISubsidiaryService.cs
@@ -2,7 +2,7 @@
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 public interface ISubsidiaryService
 {
     Task<OrganisationResponseModel?> GetCompanyByCompaniesHouseNumber(string companiesHouseNumber);
