+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/ISystemDetailsProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace EPR.SubsidiaryBulkUpload.Application.Services;
+﻿namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 public interface ISystemDetailsProvider
 {
