+++ src/EPR.SubsidiaryBulkUpload.Application/Services/Interfaces/ITableStorageProcessor.cs
@@ -1,6 +1,6 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 public interface ITableStorageProcessor
 {
