+++ src/EPR.SubsidiaryBulkUpload.Application/Services/RecordExtraction.cs
@@ -1,4 +1,5 @@
 ﻿using EPR.SubsidiaryBulkUpload.Application.DTOs;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services;
