+++ src/EPR.SubsidiaryBulkUpload.Application/Services/TableStorageProcessor.cs
@@ -2,6 +2,7 @@
 using Azure;
 using Azure.Data.Tables;
 using EPR.SubsidiaryBulkUpload.Application.Models;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 using Microsoft.Extensions.Logging;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services;
