+++ src/EPR.SubsidiaryBulkUpload.Function.UnitTests/BulkUploadFunctionTests.cs
@@ -5,7 +5,7 @@ using Azure.Storage.Blobs.Models;
 using CsvHelper.Configuration;
 using EPR.SubsidiaryBulkUpload.Application.ClassMaps;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
-using EPR.SubsidiaryBulkUpload.Application.Services;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 using EPR.SubsidiaryBulkUpload.Function.UnitTests.TestHelpers;
 using Microsoft.Extensions.Logging;
 using Moq;
