+++ src/EPR.SubsidiaryBulkUpload.Function.UnitTests/CompaniesHouseImportFunctionTests.cs
@@ -6,7 +6,7 @@ using Azure.Storage.Blobs.Models;
 using CsvHelper.Configuration;
 using EPR.SubsidiaryBulkUpload.Application.Models;
 using EPR.SubsidiaryBulkUpload.Application.Options;
-using EPR.SubsidiaryBulkUpload.Application.Services;
+using EPR.SubsidiaryBulkUpload.Application.Services.Interfaces;
 using EPR.SubsidiaryBulkUpload.Function.UnitTests.TestHelpers;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
